Make the ui loading pattern a reusable component, extract Profile VM to its own file
