+++ app/src/main/java/app/jjerrell/meta/course/sample/littlelemon/ui/composables/components/PageLoadingIndicator.kt
@@ -1,0 +1,29 @@
+package app.jjerrell.meta.course.sample.littlelemon.ui.composables.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun PageLoadingIndicator(
+    modifier: Modifier = Modifier,
+    onInitialize: () -> Unit
+) {
+    Column(
+        modifier = modifier
+            .fillMaxSize(),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center
+    ) {
+        LaunchedEffect(key1 = Unit) {
+            onInitialize()
+        }
+        CircularProgressIndicator(color = MaterialTheme.colorScheme.primary)
+    }
+}
