+++ app/src/main/java/app/jjerrell/meta/course/sample/littlelemon/ui/composables/onboarding/CheckOnboarding.kt
@@ -1,16 +1,10 @@
 package app.jjerrell.meta.course.sample.littlelemon.ui.composables.onboarding
 
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.viewmodel.compose.viewModel
+import app.jjerrell.meta.course.sample.littlelemon.ui.composables.components.PageLoadingIndicator
 
 /**
  * Checks if the user has registered or not. Calls onResult after checking status,
@@ -26,21 +20,14 @@ fun CheckOnboarding(
 ) {
     val viewModel = viewModel<CheckOnboardingViewModel>()
     val context = LocalContext.current
-    Column(
-        modifier = modifier
-            .fillMaxSize(),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        if (viewModel.isLoading) {
-            LaunchedEffect(key1 = Unit) {
+    PageLoadingIndicator(
+        modifier = modifier,
+        onInitialize = {
+            if (viewModel.isLoading) {
                 viewModel.checkRegistration(context = context)
-            }
-            CircularProgressIndicator(color = MaterialTheme.colorScheme.primary)
-        } else {
-            LaunchedEffect(key1 = Unit) {
+            } else {
                 onResult(viewModel.isRegistered)
             }
         }
-    }
+    )
 }
